+++ src/components/Main/index.jsx
@@ -1,10 +1,10 @@
-import styles from '@/components/Main.module.css';
+import styles from './Main.module.css';
 import { Links } from '@/components/Links';
 import { NextLogo } from '@/components/NextLogo';
 import { HeadDescription } from '@/components/HeadDescription';
 import { Header } from '@/components/Header';
 
-export default function Main(props) {
+export function Main(props) {
   return (
     <main className={styles.main}>
       <Header />
