+++ src/pages/_app.jsx
@@ -1,0 +1,14 @@
+import '@/styles/globals.css'
+import Head from 'next/head'
+
+export default function App({ Component, pageProps }) {
+  return (
+    <>
+      <Head>
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <Component {...pageProps} />
+    </>
+  )
+}
