+++ src/pages/index.jsx
@@ -1,0 +1,15 @@
+import Head from 'next/head';
+import { Main } from '@/components/Main';
+
+export default function Home() {
+  return (
+    <>
+      <Head>
+        <title>shun kusakabe | portfolio site</title>
+        <meta name="description" content="This is Shun's portfolio site. Generated by create next app" />
+      </Head>
+    
+      <Main title="index" />
+    </>
+  );
+}
