--- src/pages/works.js
@@ -1,17 +1,0 @@
-import Head from 'next/head';
-import Main from '@/components/Main';
-
-export default function Home() {
-  return (
-    <>
-      <Head>
-        <title>shun kusakabe | portfolio site</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <Main title="works" />
-    </>
-  );
-}
